fx-backend/requirements.txt
fastapi
uvicorn
crewai
crewai-tools
python-dotenv
pydantic
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
